+++ src/skeleton.rs
@@ -3,11 +3,11 @@ use std::{borrow::Cow, sync::Arc};
 use crate::{
     bone::Bone,
     c::{
-        spBone, spSkeleton, spSkeletonData, spSkeleton_create, spSkeleton_dispose,
+        spBone, spPhysics, spSkeleton, spSkeletonData, spSkeleton_create, spSkeleton_dispose,
         spSkeleton_getAttachmentForSlotIndex, spSkeleton_getAttachmentForSlotName,
         spSkeleton_setAttachment, spSkeleton_setBonesToSetupPose, spSkeleton_setSkin,
         spSkeleton_setSkinByName, spSkeleton_setSlotsToSetupPose, spSkeleton_setToSetupPose,
-        spSkeleton_updateCache, spSkeleton_updateWorldTransform,
+        spSkeleton_update, spSkeleton_updateCache, spSkeleton_updateWorldTransform,
         spSkeleton_updateWorldTransformWith, spSkin, spSlot,
     },
     c_interface::{to_c_str, CTmpMut, CTmpRef, NewFromPtr, SyncPtr},
@@ -15,7 +15,7 @@ use crate::{
     skeleton_data::SkeletonData,
     skin::Skin,
     slot::Slot,
-    Attachment,
+    Attachment, Physics,
 };
 
 #[allow(unused_imports)]
@@ -50,6 +50,12 @@ impl Skeleton {
         }
     }
 
+    pub fn update(&mut self, delta: f32) {
+        unsafe {
+            spSkeleton_update(self.c_ptr(), delta);
+        }
+    }
+
     /// Caches information about bones and constraints. Must be called if the skin is modified or if
     /// bones, constraints, or weighted path attachments are added or removed.
     pub fn update_cache(&mut self) {
@@ -63,9 +69,9 @@ impl Skeleton {
     /// See
     /// [`World transforms`](http://esotericsoftware.com/spine-runtime-skeletons#World-transforms)
     /// in the Spine Runtimes Guide.
-    pub fn update_world_transform(&mut self) {
+    pub fn update_world_transform(&mut self, physics: Physics) {
         unsafe {
-            spSkeleton_updateWorldTransform(self.c_ptr());
+            spSkeleton_updateWorldTransform(self.c_ptr(), physics as spPhysics);
         }
     }
 
@@ -79,8 +85,8 @@ impl Skeleton {
     /// # Safety
     ///
     /// The bone must originate from this skeleton.
-    pub unsafe fn update_world_transform_with(&mut self, parent: &Bone) {
-        spSkeleton_updateWorldTransformWith(self.c_ptr(), parent.c_ptr());
+    pub unsafe fn update_world_transform_with(&mut self, parent: &Bone, physics: Physics) {
+        spSkeleton_updateWorldTransformWith(self.c_ptr(), parent.c_ptr(), physics as spPhysics);
     }
 
     /// Sets the bones, constraints, slots, and draw order to their setup pose values.
@@ -169,7 +175,7 @@ impl Skeleton {
     /// # #[path="./test.rs"]
     /// # mod test;
     /// # use rusty_spine::{AnimationState, AnimationEvent};
-    /// # let (mut skeleton, _) = test::TestAsset::spineboy().instance();
+    /// # let (mut skeleton, _) = test::TestAsset::spineboy().instance(true);
     /// skeleton.set_skins_by_name("combined-skin", ["hat", "suit", "tie"]);
     /// ```
     ///